Use custom Firefox 52 ESR build on macOS to fix Mojave drop-down issue
https://bugzilla.mozilla.org/show_bug.cgi?id=1468053

+++ fetch_xulrunner.sh
@@ -122,7 +122,12 @@ if [ $BUILD_MAC == 1 ]; then
 	DOWNLOAD_URL="https://ftp.mozilla.org/pub/firefox/releases/$GECKO_VERSION"
 	rm -rf Firefox.app
 	
-	curl -O "$DOWNLOAD_URL/mac/en-US/Firefox%20$GECKO_VERSION.dmg"
+	if [ -e "Firefox-Modified-$GECKO_VERSION.dmg" ]; then
+		echo "Using Firefox-Modified-$GECKO_VERSION.dmg"
+		cp "Firefox-Modified-$GECKO_VERSION.dmg" "Firefox%20$GECKO_VERSION.dmg"
+	else
+		curl -O "$DOWNLOAD_URL/mac/en-US/Firefox%20$GECKO_VERSION.dmg"
+	fi
 	set +e
 	hdiutil detach -quiet /Volumes/Firefox 2>/dev/null
 	set -e
